Refactor settings page widgets
